requirement.txt
kiteconnect>=4.1.0
python-dotenv>=0.19.0
pandas>=1.3.0